Pimlico Code Redundancy Removed
