+++ 4337-gas-metering/.env.example
@@ -9,11 +9,9 @@ PIMLICO_API_KEY = "" # https://dashboard.pimlico.io/apikeys
 PIMLICO_ENTRYPOINT_ADDRESS = "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789"
 PIMLICO_MULTISEND_ADDRESS = "0x38869bf66a61cF6bDB996A6aE40D5853Fd43B526"
 PIMLICO_ERC20_PAYMASTER_ADDRESS = "0xEc43912D8C772A0Eba5a27ea5804Ba14ab502009"
-PIMLICO_USDC_TOKEN_ADDRESS = "0x07865c6E87B9F70255377e024ace6630C1Eaa37F" # https://usdcfaucet.com/
+PIMLICO_USDC_TOKEN_ADDRESS = "0x07865c6E87B9F70255377e024ace6630C1Eaa37F" # https://faucet.circle.com/
 # Make sure all nonce are unique for it to deploy account when run initially.
-PIMLICO_ACCOUNT_NONCE = "1"
-PIMLICO_ERC20_NONCE = "2"
-PIMLICO_ERC721_NONCE = "3"
+PIMLICO_NONCE = "1"
 # Pimlico Token Operation Values (Based on the chain selected, these values should also change accordingly.)
 PIMLICO_ERC20_TOKEN_CONTRACT = ""
 PIMLICO_ERC721_TOKEN_CONTRACT = ""
