+++ 4337-gas-metering/pimlico/pimlico.ts
@@ -13,27 +13,52 @@ import {
 } from "../utils/safe";
 import { submitUserOperation, signUserOperation } from "../utils/userOps";
 import { setTimeout } from "timers/promises";
-import { getERC20Decimals, getERC20Balance } from "../utils/erc20";
+import {
+  generateTransferCallData,
+  getERC20Decimals,
+  getERC20Balance,
+  mintERC20Token,
+} from "../utils/erc20";
 import { generateMintingCallData } from "../utils/erc721";
 
 dotenv.config();
 const paymaster = "pimlico";
+
 const privateKey = process.env.PRIVATE_KEY;
+
 const entryPointAddress = process.env
   .PIMLICO_ENTRYPOINT_ADDRESS as `0x${string}`;
 const multiSendAddress = process.env.PIMLICO_MULTISEND_ADDRESS as `0x${string}`;
-const saltNonce = BigInt(process.env.PIMLICO_ERC721_NONCE as string);
+
+const saltNonce = BigInt(process.env.PIMLICO_NONCE as string);
+
 const chain = process.env.PIMLICO_CHAIN;
 const chainID = Number(process.env.PIMLICO_CHAIN_ID);
+
 const safeVersion = process.env.SAFE_VERSION as string;
+
 const rpcURL = process.env.PIMLICO_RPC_URL;
 const apiKey = process.env.PIMLICO_API_KEY;
+
 const erc20PaymasterAddress = process.env
   .PIMLICO_ERC20_PAYMASTER_ADDRESS as `0x${string}`;
 const usdcTokenAddress = process.env
   .PIMLICO_USDC_TOKEN_ADDRESS as `0x${string}`;
+const erc20TokenAddress = process.env
+  .PIMLICO_ERC20_TOKEN_CONTRACT as `0x${string}`;
 const erc721TokenAddress = process.env
   .PIMLICO_ERC721_TOKEN_CONTRACT as `0x${string}`;
+
+const argv = process.argv.slice(2);
+if (argv.length != 1) {
+  throw new Error("TX Type Argument not passed."); // account || erc20 || erc721
+}
+
+const txType: string = argv[0];
+if (txType != "account" && txType != "erc20" && txType != "erc721") {
+  throw new Error("TX Type Argument Invalid");
+}
+
 const safeAddresses = (
   SAFE_ADDRESSES_MAP as Record<string, Record<string, any>>
 )[safeVersion];
@@ -107,6 +132,26 @@ console.log(
   "Address Link: https://" + chain + ".etherscan.io/address/" + senderAddress,
 );
 
+const contractCode = await publicClient.getBytecode({ address: senderAddress });
+
+if (contractCode) {
+  console.log("\nThe Safe is already deployed.");
+  if (txType == "account") {
+    console.log("");
+    process.exit(0);
+  }
+} else {
+  console.log(
+    "\nDeploying a new Safe and executing calldata passed with it (if any).\n",
+  );
+}
+
+const newNonce = await getAccountNonce(publicClient, {
+  entryPoint: entryPointAddress,
+  sender: senderAddress,
+});
+console.log("\nNonce for the sender received from EntryPoint.");
+
 // Fetch USDC balance of sender
 const usdcDecimals = await getERC20Decimals(usdcTokenAddress, publicClient);
 const usdcAmount = BigInt(10 ** usdcDecimals);
@@ -120,11 +165,11 @@ console.log(
   Number(senderUSDCBalance / usdcAmount),
 );
 
-if (senderUSDCBalance < BigInt(2) * usdcAmount) {
+if (senderUSDCBalance < BigInt(1) * usdcAmount) {
   console.log(
     "\nPlease deposit atleast 2 USDC Token for paying the Paymaster.",
   );
-  while (senderUSDCBalance < BigInt(2) * usdcAmount) {
+  while (senderUSDCBalance < BigInt(1) * usdcAmount) {
     await setTimeout(30000);
     senderUSDCBalance = await getERC20Balance(
       usdcTokenAddress,
@@ -138,42 +183,80 @@ if (senderUSDCBalance < BigInt(2) * usdcAmount) {
   );
 }
 
-const gasPriceResult = await bundlerClient.getUserOperationGasPrice();
+let txCallData!: `0x${string}`;
 
-const newNonce = await getAccountNonce(publicClient, {
-  entryPoint: entryPointAddress,
-  sender: senderAddress,
-});
-console.log("\nNonce for the sender received from EntryPoint.");
-
-const contractCode = await publicClient.getBytecode({ address: senderAddress });
-
-if (contractCode) {
-  console.log(
-    "The Safe is already deployed. Minting 1 ERC721 Token to the Safe.",
+if (txType == "account") {
+  txCallData = encodeCallData({
+    to: senderAddress,
+    data: "0x",
+    value: 0n,
+  });
+} else if (txType == "erc20") {
+  // Token Configurations
+  const erc20Decimals = await getERC20Decimals(erc20TokenAddress, publicClient);
+  const erc20Amount = BigInt(10 ** erc20Decimals);
+  let senderERC20Balance = await getERC20Balance(
+    erc20TokenAddress,
+    publicClient,
+    senderAddress,
   );
-} else {
   console.log(
-    "Deploying a new Safe and Minting 1 ERC721 Token to the Safe in one tx",
+    "\nSafe Wallet ERC20 Balance:",
+    Number(senderERC20Balance / erc20Amount),
   );
+
+  // Trying to mint tokens (Make sure ERC20 Token Contract is mintable by anyone).
+  if (senderERC20Balance < erc20Amount) {
+    console.log("\nMinting ERC20 Tokens to Safe Wallet.");
+    await mintERC20Token(
+      erc20TokenAddress,
+      publicClient,
+      signer,
+      senderAddress,
+      erc20Amount,
+      chain,
+      paymaster,
+    );
+
+    while (senderERC20Balance < erc20Amount) {
+      await setTimeout(15000);
+      senderERC20Balance = await getERC20Balance(
+        erc20TokenAddress,
+        publicClient,
+        senderAddress,
+      );
+    }
+    console.log(
+      "\nUpdated Safe Wallet ERC20 Balance:",
+      Number(senderERC20Balance / erc20Amount),
+    );
+  }
+  txCallData = encodeCallData({
+    to: erc20TokenAddress,
+    data: generateTransferCallData(signer.address, erc20Amount), // transfer() function call with corresponding data.
+    value: 0n,
+  });
+} else if (txType == "erc721") {
+  txCallData = encodeCallData({
+    to: erc721TokenAddress,
+    data: generateMintingCallData(signer.address), // safeMint() function call with corresponding data.
+    value: 0n,
+  });
 }
+
+const gasPriceResult = await bundlerClient.getUserOperationGasPrice();
 
 const sponsoredUserOperation: UserOperation = {
   sender: senderAddress,
   nonce: newNonce,
   initCode: contractCode ? "0x" : initCode,
-  // Minting 1 ERC721 Token to the Safe itself
-  callData: encodeCallData({
-    to: erc721TokenAddress,
-    data: generateMintingCallData(senderAddress),
-    value: 0n,
-  }),
+  callData: txCallData,
   callGasLimit: 100_000n, // Gas Values Hardcoded for now at a high value
   verificationGasLimit: 500_000n,
   preVerificationGas: 50_000n,
   maxFeePerGas: gasPriceResult.fast.maxFeePerGas,
   maxPriorityFeePerGas: gasPriceResult.fast.maxPriorityFeePerGas,
-  paymasterAndData: erc20PaymasterAddress, // to use the erc20 paymaster, put its address in the paymasterAndData field
+  paymasterAndData: erc20PaymasterAddress,
   signature: "0x",
 };
